White Flag
While I would love to continue work and figure this all out, time has
run out on me. Will record as is to the best of my ability. I apologize
that I was unable to complete the project, and apologize that grading
will be harder on you than if everything worked perfectly.